+++ earthaccess/store.py
@@ -64,12 +64,15 @@ def _open_files(
     url_mapping: Mapping[str, Union[DataGranule, None]],
     fs: fsspec.AbstractFileSystem,
     threads: Optional[int] = 8,
+    pqdm_kwargs: Optional[Mapping[str, Any]] = None,
 ) -> List[EarthAccessFile]:
     def multi_thread_open(data: tuple) -> EarthAccessFile:
         urls, granule = data
         return EarthAccessFile(fs.open(urls), granule)
 
-    fileset = pqdm(url_mapping.items(), multi_thread_open, n_jobs=threads)
+    fileset = pqdm(
+        url_mapping.items(), multi_thread_open, n_jobs=threads, **pqdm_kwargs
+    )
     return fileset
 
 
@@ -336,6 +339,7 @@ class Store(object):
         self,
         granules: Union[List[str], List[DataGranule]],
         provider: Optional[str] = None,
+        pqdm_kwargs: Optional[Mapping[str, Any]] = None,
     ) -> List[EarthAccessFile]:
         """Returns a list of file-like objects that can be used to access files
         hosted on S3 or HTTPS by third party libraries like xarray.
@@ -349,7 +353,7 @@ class Store(object):
             A list of "file pointers" to remote (i.e. s3 or https) files.
         """
         if len(granules):
-            return self._open(granules, provider)
+            return self._open(granules, provider, pqdm_kwargs)
         return []
 
     @singledispatchmethod
@@ -357,6 +361,7 @@ class Store(object):
         self,
         granules: Union[List[str], List[DataGranule]],
         provider: Optional[str] = None,
+        pqdm_kwargs: Optional[Mapping[str, Any]] = None,
     ) -> List[Any]:
         raise NotImplementedError("granules should be a list of DataGranule or URLs")
 
@@ -366,6 +371,7 @@ class Store(object):
         granules: List[DataGranule],
         provider: Optional[str] = None,
         threads: Optional[int] = 8,
+        pqdm_kwargs: Optional[Mapping[str, Any]] = None,
     ) -> List[Any]:
         fileset: List = []
         total_size = round(sum([granule.size() for granule in granules]) / 1024, 2)
@@ -391,14 +397,22 @@ class Store(object):
             else:
                 access = "on_prem"
                 s3_fs = None
+                access = "direct"
+                provider = granules[0]["meta"]["provider-id"]
+                # if the data has its own S3 credentials endpoint, we will use it
+                endpoint = self._own_s3_credentials(granules[0]["umm"]["RelatedUrls"])
+                if endpoint is not None:
+                    logger.info(f"using endpoint: {endpoint}")
+                    s3_fs = self.get_s3_filesystem(endpoint=endpoint)
+                else:
+                    logger.info(f"using provider: {provider}")
+                    s3_fs = self.get_s3_filesystem(provider=provider)
 
             url_mapping = _get_url_granule_mapping(granules, access)
             if s3_fs is not None:
                 try:
                     fileset = _open_files(
-                        url_mapping,
-                        fs=s3_fs,
-                        threads=threads,
+                        url_mapping, fs=s3_fs, threads=threads, pqdm_kwargs=pqdm_kwargs
                     )
                 except Exception as e:
                     raise RuntimeError(
@@ -407,11 +421,15 @@ class Store(object):
                         f"Exception: {traceback.format_exc()}"
                     ) from e
             else:
-                fileset = self._open_urls_https(url_mapping, threads=threads)
+                fileset = self._open_urls_https(
+                    url_mapping, threads=threads, pqdm_kwargs=pqdm_kwargs
+                )
             return fileset
         else:
             url_mapping = _get_url_granule_mapping(granules, access="on_prem")
-            fileset = self._open_urls_https(url_mapping, threads=threads)
+            fileset = self._open_urls_https(
+                url_mapping, threads=threads, pqdm_kwargs=pqdm_kwargs
+            )
             return fileset
 
     @_open.register
@@ -420,6 +438,7 @@ class Store(object):
         granules: List[str],
         provider: Optional[str] = None,
         threads: Optional[int] = 8,
+        pqdm_kwargs: Optional[Mapping[str, Any]] = None,
     ) -> List[Any]:
         fileset: List = []
 
@@ -447,6 +466,7 @@ class Store(object):
                             url_mapping,
                             fs=s3_fs,
                             threads=threads,
+                            pqdm_kwargs=pqdm_kwargs,
                         )
                     except Exception as e:
                         raise RuntimeError(
@@ -466,7 +486,7 @@ class Store(object):
                 raise ValueError(
                     "We cannot open S3 links when we are not in-region, try using HTTPS links"
                 )
-            fileset = self._open_urls_https(url_mapping, threads)
+            fileset = self._open_urls_https(url_mapping, threads, pqdm_kwargs)
             return fileset
 
     def get(
@@ -475,6 +495,7 @@ class Store(object):
         local_path: Union[Path, str, None] = None,
         provider: Optional[str] = None,
         threads: int = 8,
+        pqdm_kwargs: Optional[Mapping[str, Any]] = None,
     ) -> List[str]:
         """Retrieves data granules from a remote storage system.
 
@@ -503,7 +524,7 @@ class Store(object):
             local_path = Path(local_path)
 
         if len(granules):
-            files = self._get(granules, local_path, provider, threads)
+            files = self._get(granules, local_path, provider, threads, pqdm_kwargs)
             return files
         else:
             raise ValueError("List of URLs or DataGranule instances expected")
@@ -515,6 +536,7 @@ class Store(object):
         local_path: Path,
         provider: Optional[str] = None,
         threads: int = 8,
+        pqdm_kwargs: Optional[Mapping[str, Any]] = None,
     ) -> List[str]:
         """Retrieves data granules from a remote storage system.
 
@@ -544,6 +566,7 @@ class Store(object):
         local_path: Path,
         provider: Optional[str] = None,
         threads: int = 8,
+        pqdm_kwargs: Optional[Mapping[str, Any]] = None,
     ) -> List[str]:
         data_links = granules
         downloaded_files: List = []
@@ -565,7 +588,9 @@ class Store(object):
 
         else:
             # if we are not in AWS
-            return self._download_onprem_granules(data_links, local_path, threads)
+            return self._download_onprem_granules(
+                data_links, local_path, threads, pqdm_kwargs
+            )
 
     @_get.register
     def _get_granules(
@@ -574,6 +599,7 @@ class Store(object):
         local_path: Path,
         provider: Optional[str] = None,
         threads: int = 8,
+        pqdm_kwargs: Optional[Mapping[str, Any]] = None,
     ) -> List[str]:
         data_links: List = []
         downloaded_files: List = []
@@ -614,7 +640,9 @@ class Store(object):
         else:
             # if the data are cloud-based, but we are not in AWS,
             # it will be downloaded as if it was on prem
-            return self._download_onprem_granules(data_links, local_path, threads)
+            return self._download_onprem_granules(
+                data_links, local_path, threads, pqdm_kwargs
+            )
 
     def _download_file(self, url: str, directory: Path) -> str:
         """Download a single file from an on-prem location, a DAAC data center.
@@ -652,7 +680,11 @@ class Store(object):
         return str(path)
 
     def _download_onprem_granules(
-        self, urls: List[str], directory: Path, threads: int = 8
+        self,
+        urls: List[str],
+        directory: Path,
+        threads: int = 8,
+        pqdm_kwargs: Optional[Mapping[str, Any]] = None,
     ) -> List[Any]:
         """Downloads a list of URLS into the data directory.
 
@@ -661,6 +693,9 @@ class Store(object):
             directory: local directory to store the downloaded files
             threads: parallel number of threads to use to download the files;
                 adjust as necessary, default = 8
+            fail_fast: if set to True, the download process will stop immediately
+            upon encountering the first error. If set to False, errors will be
+            deferred, allowing the download of remaining files to continue.
 
         Returns:
             A list of local filepaths to which the files were downloaded.
@@ -674,11 +709,13 @@ class Store(object):
         directory.mkdir(parents=True, exist_ok=True)
 
         arguments = [(url, directory) for url in urls]
+
         results = pqdm(
             arguments,
             self._download_file,
             n_jobs=threads,
             argument_type="args",
+            **pqdm_kwargs,
         )
         return results
 
@@ -686,11 +723,12 @@ class Store(object):
         self,
         url_mapping: Mapping[str, Union[DataGranule, None]],
         threads: Optional[int] = 8,
+        pqdm_kwargs: Optional[Mapping[str, Any]] = None,
     ) -> List[fsspec.AbstractFileSystem]:
         https_fs = self.get_fsspec_session()
         if https_fs is not None:
             try:
-                fileset = _open_files(url_mapping, https_fs, threads)
+                fileset = _open_files(url_mapping, https_fs, threads, pqdm_kwargs)
             except Exception:
                 logger.exception(
                     "An exception occurred while trying to access remote files via HTTPS"